+++ src/Core/global_functions.php
@@ -24,6 +24,17 @@ if (!function_exists('isPositive')) {
     }
 }
 
+if (!function_exists('isWin')) {
+    /**
+     * Returns `true` if the environment is Windows
+     * @return bool
+     */
+    function isWin()
+    {
+        return DS === '\\';
+    }
+}
+
 if (!function_exists('rtr')) {
     /**
      * Returns the relative path (to the APP root) of an absolute path
@@ -44,13 +55,12 @@ if (!function_exists('which')) {
      */
     function which($command)
     {
-        $isWin = DS == '\\';
-        $which = $isWin ? 'where' : 'which';
+        $executable = isWin() ? 'where' : 'which';
 
-        exec(sprintf('%s %s 2>&1', $which, $command), $path, $exitCode);
+        exec(sprintf('%s %s 2>&1', $executable, $command), $path, $exitCode);
 
-        $path = $isWin && !empty($path) ? array_map('escapeshellarg', $path) : $path;
+        $path = isWin() && !empty($path) ? array_map('escapeshellarg', $path) : $path;
 
-        return $exitCode === 0 && !empty($path[0]) ? $path[0] : null;
+        return $exitCode === 0 && !empty($path[0]) ? escapeshellarg($path[0]) : null;
     }
 }